split accessor into separate file
